Pt_1_4: Full task
I wrote a program inviting the user to enter a number from 10 to 20.
If the entered value is less than 10, print the message
"The number is less than the required range" and offer to try again.
If the entered value is greater than 20, the message
"The number is greater than the required range" is displayed and
the user is prompted to try again. Repeat until a value from
the range of 10 to 20 is entered, after which
the "Thank you" message is displayed.