Pt_2_2: Full task
I have written a program that works like this:
Create a list with the names of four television programs and display
them on separate lines. Invite the user to enter the name of one more
program and the position at which where it should be inserted into the
list. Display the list again,
with all five programs at the new positions.