Pt_2_5: Full task
I have written a program that plays heads or tails with the user.
The program should ask the user: heads or tails. If the user enters 0,
he chooses heads, 1 - tails, any other number - the end of the game.
The program keeps a record of wins and losses after each round. After
three consecutive losses the game ends.