Pt_5_2: Full task
I wrote a program that works like this:
The user is asked how many entries he wants to add to the list,
then the user enters entries. After the data is added, the user
is asked for the author and all books of the specified author from
the list are displayed. If there are no books by this author in the
list, the corresponding message is displayed.