Pt_5_3: Full task
I have written a program that works as follows:
Using the file books.csv, the program prompts the user to enter the
beginning and end year. It then displays all the books published in
a given period of time. The program's behavior in case of an invalid
request is provided for.