Pt_6_1: Full task
I have written a parser program to extract the top 10 most popular
songs of your favorite artist from a web resource (according to the
version of your chosen service). The user inputs a link to the
selected artist on the given web resource. In case of incorrectly
specified link or other scenario of incorrect operation error
handling is provided